requirements.txt
pylint
pytest
pytest-env
pytest-cov
requests-mock
moto